requirements.txt
absl_py>=0.8.1
imageio>=2.6.1
matplotlib>=3.0.3
monty>=3.0.2
numpy>=1.16.2
Pillow>=6.2.1
scikit-learn~=0.23.0
scipy>=1.2.1
dm_sonnet==1.35
tensorflow==1.15.0
tensorflow-gpu==1.15.0
tensorflow_probability==0.8.0
tensorflow_datasets==1.3.0
tqdm~=4.46.0
joblib~=0.16.0